requirements.txt
tarantool~=0.9.0
Flask~=2.1.2
Pillow~=9.2.0
image-similarity-measure==0.3.5
python-Levenshtein~=0.12.2
opencv-python~=4.6.0.66